pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Gabriel Lima de Souza", email = "gabriellima2803@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

